refactor(desktop): extract the vault target details block
